refactor: move ViewModels to SnapX.CommonUI
