+++ SnapX.Avalonia/App.axaml.cs
@@ -5,11 +5,8 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Layout;
 using Avalonia.Markup.Xaml;
 using Avalonia.Media;
-using Avalonia.Media.Imaging;
-using Avalonia.Platform;
 using Avalonia.Styling;
 using CommunityToolkit.Mvvm.DependencyInjection;
-using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
 using FluentAvalonia.UI.Windowing;
 using Microsoft.Extensions.DependencyInjection;
@@ -18,9 +15,8 @@ using SnapX.Avalonia.ViewModels;
 using SnapX.Avalonia.Views;
 using SnapX.Avalonia.Views.Settings;
 using SnapX.Avalonia.Views.Settings.Views;
+using SnapX.CommonUI.ViewModels;
 using SnapX.Core;
-using SnapX.Core.Capture;
-using SnapX.Core.Job;
 using SnapX.Core.Utils;
 using SnapX.Core.Utils.Native;
 
@@ -215,7 +211,7 @@ public partial class App : Application
         Console.WriteLine("Report Error button clicked. No action yet. If you have telemetry enabled, (it is by default) it will have already been sent to Sentry.");
     }
 
-    private void Shutdown()
+    public static void Shutdown()
     {
         try
         {
@@ -253,7 +249,7 @@ public partial class App : Application
                         DebugHelper.WriteLine("Received Shutdown from Avalonia");
                         if (sigintReceived) return;
                         sigintReceived = true;
-                        SnapX.shutdown();
+                        Shutdown();
 
                         // desktop.Shutdown();
                     };
@@ -264,22 +260,9 @@ public partial class App : Application
                         if (sigintReceived) return;
                         ea.Cancel = true;
                         sigintReceived = true;
-                        SnapX.shutdown();
+                        Shutdown();
                         desktop.Shutdown();
                     };
-                    // AppDomain.CurrentDomain.ProcessExit += (o, _) =>
-                    // {
-                    //     if (!sigintReceived)
-                    //     {
-                    //         sigintReceived = true;
-                    //         DebugHelper.WriteLine("Received SIGTERM");
-                    //         SnapX.shutdown();
-                    //     }
-                    //     else
-                    //     {
-                    //         DebugHelper.WriteLine("Received SIGTERM, ignoring it because already processed SIGINT");
-                    //     }
-                    // };
                     var errorStarting = false;
                     // DebugHelper.Logger.Debug($"Avalonia Args: {desktop.Args}");
                     try
@@ -298,97 +281,6 @@ public partial class App : Application
 
                     if (errorStarting) return;
                     DebugHelper.WriteLine("Internal Startup time: {0} ms", SnapX.getStartupTime());
-
-                    var logoBitmap = new Bitmap(AssetLoader.Open(new Uri("avares://snapx-ui/SnapX_Logo.png")));
-                    var trayIcon = new TrayIcon
-                    {
-                        Icon = new WindowIcon(logoBitmap),
-                        ToolTipText = Core.SnapX.AppName,
-                        Command = OpenSnapXCommand
-                    };
-
-                    var menu = new NativeMenu();
-                    menu.Opening += NativeMenu_OnOpening;
-                    menu.NeedsUpdate += NativeMenu_OnNeedsUpdate;
-
-                    var about = new NativeMenuItem(TrayTitle)
-                    {
-                        Icon = logoBitmap,
-                        ToolTip = Lang.AboutSnapX
-                    };
-                    about.Click += NativeMenuItem_SnapX_OnClick;
-                    menu.Items.Add(about);
-                    menu.Items.Add(new NativeMenuItemSeparator());
-
-                    var capture = new NativeMenuItem("Capture") { Menu = new NativeMenu() };
-                    var full = new NativeMenuItem(Lang.UI_Capture_Fullscreen);
-                    full.Click += NativeMenuItem_Capture_Fullscreen_OnClick;
-                    capture.Menu.Items.Add(full);
-
-                    var windowPicker = new NativeMenuItem(Lang.UI_Dropdown_Window)
-                    {
-                        Menu = new NativeMenu
-                    {
-                        new NativeMenuItem("SnapX UI") { Icon = logoBitmap },
-                        new NativeMenuItem("Marvel Rivals") { Icon = logoBitmap },
-                        new NativeMenuItem("Man of Steel (2013)") { Icon = logoBitmap }
-                    }
-                    };
-                    capture.Menu.Items.Add(windowPicker);
-
-                    var monitorPicker = new NativeMenuItem(Lang.UI_Dropdown_Monitor) { Menu = [] };
-                    monitorPicker.Menu.NeedsUpdate += NativeMenu_OnNeedsUpdate;
-                    capture.Menu.Items.Add(monitorPicker);
-
-                    capture.Menu.Items.Add(new NativeMenuItem("Region"));
-                    capture.Menu.Items.Add(new NativeMenuItem("Region (Light)"));
-                    capture.Menu.Items.Add(new NativeMenuItem("Region (Transparent)"));
-                    menu.Items.Add(capture);
-
-                    menu.Items.Add(new NativeMenuItem("Upload")
-                    {
-                        Menu = new NativeMenu
-                    {
-                        new NativeMenuItem("Upload File..."),
-                        new NativeMenuItem("Upload Folder..."),
-                        new NativeMenuItem("Upload from clipboard..."),
-                        new NativeMenuItem("Upload text..."),
-                        new NativeMenuItem("Drag and drop upload..."),
-                        new NativeMenuItem("Shorten URL..."),
-                        new NativeMenuItem("Tweet message...")
-                    }
-                    });
-                    var captureFullscreenMenuItem = new NativeMenuItem("Capture entire screen");
-                    captureFullscreenMenuItem.Click += NativeMenuItem_Capture_Fullscreen_OnClick;
-                    var captureActiveWindowMenuItem = new NativeMenuItem("Capture active window");
-                    captureActiveWindowMenuItem.Click += NativeMenuItem_Workflows_CaptureActiveWindow_OnClick;
-                    var captureActiveScreenMenuItem = new NativeMenuItem("Capture active screen");
-                    captureActiveScreenMenuItem.Click += NativeMenuItem_Workflows_CaptureActiveScreen_OnClick;
-                    var workflows = new NativeMenuItem("Workflows")
-                    {
-                        Menu =
-                        [
-                            captureFullscreenMenuItem,
-                        captureActiveScreenMenuItem,
-                        captureActiveWindowMenuItem,
-                    ]
-                    };
-
-                    menu.Items.Add(workflows);
-
-                    menu.Items.Add(new NativeMenuItemSeparator());
-
-                    var open = new NativeMenuItem("Open");
-                    open.Command = OpenSnapXCommand;
-                    menu.Items.Add(open);
-
-                    var quit = new NativeMenuItem("Quit");
-                    quit.Click += NativeMenuItem_Quit_OnClick;
-                    menu.Items.Add(quit);
-
-                    trayIcon.Menu = menu;
-
-                    TrayIcon.SetIcons(Current, [trayIcon]);
 
                     if (SnapX.isSilent()) return;
                     if (SnapX.GetCLIManager().IsCommandExist("video"))
@@ -452,6 +344,10 @@ public partial class App : Application
             aboutWindow.Show();
         }
     }
+    public static void NativeMenuAboutSnapXClick(object? Sender, EventArgs E)
+    {
+        CreateAboutWindowStatic();
+    }
     public static void CreateSettingsWindowStatic()
     {
         var settingsWindow = Design.IsDesignMode
@@ -469,12 +365,6 @@ public partial class App : Application
             settingsWindow.ShowAsDialog = false;
             settingsWindow.Show();
         }
-    }
-
-
-    private void NativeMenuAboutSnapXClick(object? Sender, EventArgs E)
-    {
-        CreateAboutWindowStatic();
     }
 
     public static void ConfigureServices(IServiceCollection services)
@@ -498,59 +388,5 @@ public partial class App : Application
         services.AddSingleton<HomePageViewModel>();
 
         services.AddSingleton<IMessenger>(WeakReferenceMessenger.Default);
-    }
-
-    private void NativeMenuItem_Quit_OnClick(object? Sender, EventArgs E)
-    {
-        Shutdown();
-    }
-
-    private void NativeMenuItem_SnapX_OnClick(object? Sender, EventArgs E)
-    {
-        NativeMenuAboutSnapXClick(Sender, E);
-    }
-
-    private void NativeMenuItem_Capture_Fullscreen_OnClick(object? Sender, EventArgs E)
-    {
-        TaskHelpers.GetScreenshot().CaptureFullscreen();
-    }
-
-    private void NativeMenuItem_Workflows_CaptureActiveScreen_OnClick(object? Sender, EventArgs E)
-    {
-        new CaptureActiveMonitor().Capture(TaskSettings.GetDefaultTaskSettings());
-    }
-
-    private void NativeMenuItem_Workflows_CaptureActiveWindow_OnClick(object? Sender, EventArgs E)
-    {
-        new CaptureActiveWindow().Capture(TaskSettings.GetDefaultTaskSettings());
-    }
-
-    private void NativeMenuItem_Open_OnClick(object? Sender, EventArgs E)
-    {
-        if (!MyMainWindow?.IsLoaded ?? false)
-        {
-            var vm = Ioc.Default.GetRequiredService<MainViewModel>();
-            MyMainWindow = new MainWindow(vm);
-            MyMainWindow.Show();
-        }
-
-        if (!MyMainWindow?.IsVisible ?? true) MyMainWindow?.Show();
-        MyMainWindow?.Focus();
-        MyMainWindow?.Activate();
-    }
-
-    [RelayCommand]
-    private void OpenSnapX()
-    {
-        NativeMenuItem_Open_OnClick(this, EventArgs.Empty);
-    }
-    private void NativeMenu_OnNeedsUpdate(object? Sender, EventArgs E)
-    {
-        DebugHelper.WriteLine("NativeMenu_OnNeedsUpdate");
-    }
-
-    private void NativeMenu_OnOpening(object? Sender, EventArgs E)
-    {
-        DebugHelper.WriteLine("NativeMenu_OnOpening");
     }
 }
