+++ SnapX.Avalonia/Views/HomePageView.axaml.cs
@@ -2,9 +2,9 @@ using AsyncImageLoader.Loaders;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 using FluentAvalonia.UI.Controls;
-using SnapX.Avalonia.Models;
 using SnapX.Avalonia.Services;
 using SnapX.Avalonia.ViewModels;
+using SnapX.CommonUI.Models;
 using SnapX.Core;
 using SnapX.Core.Utils;
 using SnapX.Core.Utils.Miscellaneous;
