+++ SnapX.Avalonia/Views/MainView.axaml.cs
@@ -3,7 +3,7 @@ using Avalonia.Input;
 using Avalonia.LogicalTree;
 using CommunityToolkit.Mvvm.Input;
 using FluentAvalonia.UI.Controls;
-using SnapX.Avalonia.ViewModels;
+using SnapX.CommonUI.ViewModels;
 using SnapX.Core;
 using SnapX.Core.Job;
 using SnapX.Core.Upload;
