+++ SnapX.Avalonia/Views/OCR.axaml
@@ -6,7 +6,7 @@
     x:Class="SnapX.Avalonia.Views.OCR"
     x:DataType="viewModels:OCRViewModel"
     xmlns="https://github.com/avaloniaui"
-    xmlns:viewModels="clr-namespace:SnapX.Avalonia.ViewModels"
+    xmlns:viewModels="clr-namespace:SnapX.CommonUI.ViewModels;assembly=SnapX.CommonUI"
     xmlns:windowing="clr-namespace:FluentAvalonia.UI.Windowing;assembly=FluentAvalonia"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
