+++ SnapX.Avalonia/Views/RegionSelectorWindow.axaml
@@ -24,7 +24,7 @@
     xmlns="https://github.com/avaloniaui"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewModels="clr-namespace:SnapX.Avalonia.ViewModels"
+    xmlns:viewModels="clr-namespace:SnapX.CommonUI.ViewModels;assembly=SnapX.CommonUI"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <Design.DataContext>
         <viewModels:RegionSelectorViewModel />
