+++ SnapX.CommonUI/Models/ListItemTemplate.cs
@@ -1,3 +1,3 @@
-﻿namespace SnapX.Avalonia.Models;
+﻿namespace SnapX.CommonUI.Models;
 
 public record ListItemTemplate(Type ModelType, string IconKey, string Label);
