+++ SnapX.CommonUI/Models/ListTaskTemplate.cs
@@ -1,5 +1,5 @@
 ﻿using SnapX.Core.History;
 
-namespace SnapX.Avalonia.Models;
+namespace SnapX.CommonUI.Models;
 
 public record ListTaskTemplate(Type ModelType, HistoryItem task);
