--- SnapX.Core/Models/SavedConfiguration.cs
@@ -1,22 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-
-namespace SnapX.Core.Models;
-
-[Table("ApplicationConfig")]
-public record SavedConfiguration
-{
-    [Key]
-    public int Id { get; set; }
-
-    public string ConfigSection { get; set; }
-
-    public string SettingKey { get; set; }
-
-    public string SettingValue { get; set; }
-
-    public string DataType { get; set; }
-
-    public DateTime UpdatedAt { get; set; }
-}
